cmd/trace: Move TraceParser to a common package
