+++ cmd/kubectl-gadget/trace/bind.go
@@ -19,6 +19,7 @@ import (
 	"strconv"
 	"strings"
 
+	commontrace "github.com/kinvolk/inspektor-gadget/cmd/common/trace"
 	commonutils "github.com/kinvolk/inspektor-gadget/cmd/common/utils"
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/trace/bind/types"
@@ -108,7 +109,7 @@ func newBindCmd() *cobra.Command {
 	return cmd
 }
 
-func NewBindParser(outputConfig *commonutils.OutputConfig) TraceParser[types.Event] {
+func NewBindParser(outputConfig *commonutils.OutputConfig) commontrace.TraceParser[types.Event] {
 	columnsWidth := map[string]int{
 		"node":      -16,
 		"namespace": -16,
