+++ cmd/kubectl-gadget/trace/dns.go
@@ -18,6 +18,7 @@ import (
 	"fmt"
 	"strings"
 
+	commontrace "github.com/kinvolk/inspektor-gadget/cmd/common/trace"
 	commonutils "github.com/kinvolk/inspektor-gadget/cmd/common/utils"
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/trace/dns/types"
@@ -64,7 +65,7 @@ func newDNSCmd() *cobra.Command {
 	return cmd
 }
 
-func NewDNSParser(outputConfig *commonutils.OutputConfig) TraceParser[types.Event] {
+func NewDNSParser(outputConfig *commonutils.OutputConfig) commontrace.TraceParser[types.Event] {
 	columnsWidth := map[string]int{
 		"node":      -16,
 		"namespace": -16,
