+++ cmd/kubectl-gadget/trace/exec.go
@@ -18,6 +18,7 @@ import (
 	"fmt"
 	"strings"
 
+	commontrace "github.com/kinvolk/inspektor-gadget/cmd/common/trace"
 	commonutils "github.com/kinvolk/inspektor-gadget/cmd/common/utils"
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/trace/exec/types"
@@ -67,7 +68,7 @@ func newExecCmd() *cobra.Command {
 	return cmd
 }
 
-func NewExecParser(outputConfig *commonutils.OutputConfig) TraceParser[types.Event] {
+func NewExecParser(outputConfig *commonutils.OutputConfig) commontrace.TraceParser[types.Event] {
 	columnsWidth := map[string]int{
 		"node":      -16,
 		"namespace": -16,
