+++ cmd/kubectl-gadget/trace/fsslower.go
@@ -20,6 +20,7 @@ import (
 	"strconv"
 	"strings"
 
+	commontrace "github.com/kinvolk/inspektor-gadget/cmd/common/trace"
 	commonutils "github.com/kinvolk/inspektor-gadget/cmd/common/utils"
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/trace/fsslower/types"
@@ -112,7 +113,7 @@ func newFsSlowerCmd() *cobra.Command {
 	return cmd
 }
 
-func NewFsslowerParser(outputConfig *commonutils.OutputConfig) TraceParser[types.Event] {
+func NewFsslowerParser(outputConfig *commonutils.OutputConfig) commontrace.TraceParser[types.Event] {
 	columnsWidth := map[string]int{
 		"node":      -16,
 		"namespace": -16,
