+++ cmd/kubectl-gadget/trace/mount.go
@@ -18,6 +18,7 @@ import (
 	"fmt"
 	"strings"
 
+	commontrace "github.com/kinvolk/inspektor-gadget/cmd/common/trace"
 	commonutils "github.com/kinvolk/inspektor-gadget/cmd/common/utils"
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/trace/mount/types"
@@ -67,7 +68,7 @@ func newMountCmd() *cobra.Command {
 	return cmd
 }
 
-func NewMountParser(outputConfig *commonutils.OutputConfig) TraceParser[types.Event] {
+func NewMountParser(outputConfig *commonutils.OutputConfig) commontrace.TraceParser[types.Event] {
 	columnsWidth := map[string]int{
 		"node":      -16,
 		"namespace": -16,
