+++ cmd/kubectl-gadget/trace/network.go
@@ -18,6 +18,7 @@ import (
 	"fmt"
 	"strings"
 
+	commontrace "github.com/kinvolk/inspektor-gadget/cmd/common/trace"
 	commonutils "github.com/kinvolk/inspektor-gadget/cmd/common/utils"
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/trace/network/types"
@@ -65,7 +66,7 @@ func newNetworkCmd() *cobra.Command {
 	return cmd
 }
 
-func NewNetworkParser(outputConfig *commonutils.OutputConfig) TraceParser[types.Event] {
+func NewNetworkParser(outputConfig *commonutils.OutputConfig) commontrace.TraceParser[types.Event] {
 	columnsWidth := map[string]int{
 		"node":      -16,
 		"namespace": -16,
