+++ cmd/kubectl-gadget/trace/oomkill.go
@@ -20,6 +20,7 @@ import (
 
 	"github.com/spf13/cobra"
 
+	commontrace "github.com/kinvolk/inspektor-gadget/cmd/common/trace"
 	commonutils "github.com/kinvolk/inspektor-gadget/cmd/common/utils"
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/trace/oomkill/types"
@@ -67,7 +68,7 @@ func newOOMKillCmd() *cobra.Command {
 	return cmd
 }
 
-func NewOOMKillParser(outputConfig *commonutils.OutputConfig) TraceParser[types.Event] {
+func NewOOMKillParser(outputConfig *commonutils.OutputConfig) commontrace.TraceParser[types.Event] {
 	columnsWidth := map[string]int{
 		"node":      -16,
 		"namespace": -16,
