+++ cmd/kubectl-gadget/trace/open.go
@@ -18,6 +18,7 @@ import (
 	"fmt"
 	"strings"
 
+	commontrace "github.com/kinvolk/inspektor-gadget/cmd/common/trace"
 	commonutils "github.com/kinvolk/inspektor-gadget/cmd/common/utils"
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/trace/open/types"
@@ -67,7 +68,7 @@ func newOpenCmd() *cobra.Command {
 	return cmd
 }
 
-func NewOpenParser(outputConfig *commonutils.OutputConfig) TraceParser[types.Event] {
+func NewOpenParser(outputConfig *commonutils.OutputConfig) commontrace.TraceParser[types.Event] {
 	columnsWidth := map[string]int{
 		"node":      -16,
 		"namespace": -16,
