+++ cmd/kubectl-gadget/trace/signal.go
@@ -19,6 +19,7 @@ import (
 	"strconv"
 	"strings"
 
+	commontrace "github.com/kinvolk/inspektor-gadget/cmd/common/trace"
 	commonutils "github.com/kinvolk/inspektor-gadget/cmd/common/utils"
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/trace/signal/types"
@@ -102,7 +103,7 @@ func newSignalCmd() *cobra.Command {
 	return cmd
 }
 
-func NewSignalParser(outputConfig *commonutils.OutputConfig) TraceParser[types.Event] {
+func NewSignalParser(outputConfig *commonutils.OutputConfig) commontrace.TraceParser[types.Event] {
 	columnsWidth := map[string]int{
 		"node":      -16,
 		"namespace": -16,
