+++ cmd/kubectl-gadget/trace/tcp.go
@@ -18,6 +18,7 @@ import (
 	"fmt"
 	"strings"
 
+	commontrace "github.com/kinvolk/inspektor-gadget/cmd/common/trace"
 	commonutils "github.com/kinvolk/inspektor-gadget/cmd/common/utils"
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/trace/tcp/types"
@@ -70,7 +71,7 @@ func newTCPCmd() *cobra.Command {
 	return cmd
 }
 
-func NewTCPParser(outputConfig *commonutils.OutputConfig) TraceParser[types.Event] {
+func NewTCPParser(outputConfig *commonutils.OutputConfig) commontrace.TraceParser[types.Event] {
 	columnsWidth := map[string]int{
 		"node":      -16,
 		"namespace": -16,
