+++ cmd/kubectl-gadget/trace/tcpconnect.go
@@ -18,6 +18,7 @@ import (
 	"fmt"
 	"strings"
 
+	commontrace "github.com/kinvolk/inspektor-gadget/cmd/common/trace"
 	commonutils "github.com/kinvolk/inspektor-gadget/cmd/common/utils"
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/utils"
 	"github.com/kinvolk/inspektor-gadget/pkg/gadgets/trace/tcpconnect/types"
@@ -68,7 +69,7 @@ func newTcpconnectCmd() *cobra.Command {
 	return cmd
 }
 
-func NewTcpconnectParser(outputConfig *commonutils.OutputConfig) TraceParser[types.Event] {
+func NewTcpconnectParser(outputConfig *commonutils.OutputConfig) commontrace.TraceParser[types.Event] {
 	columnsWidth := map[string]int{
 		"node":      -16,
 		"namespace": -16,
